Connectly/requirements.txt
pandas>=2.3.1
sqlalchemy>=2.0.41
mysql-connector-python>=9.3.0
networkx>=3.5
flask>=3.1.1
flask-cors>=6.0.1
flask-socketio==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2
pyvis>=0.3.2
matplotlib>=3.5.0
scipy>=1.10.1
uuid>=1.30
gunicorn>=21.2.0
python-dotenv>=1.0.0
eventlet>=0.33.3
